# tests/__snapshots__/openapi.core.json
{
  "components": {
    "schemas": {
...
      "AgentPackEnvelopeDto": {
        "properties": {
          "integrity": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentPackIntegrityDto"
              }
            ]
          },
          "manifest": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentPackManifestDto"
              }
            ]
          }
        },
        "type": "object"
      },
...
      "AgentPackPreviewView": {
        "properties": {
          "action": {
            "type": "string"
          },
          "bundled_version": {
            "type": "string"
          },
          "counts": {
            "$ref": "#/components/schemas/AgentPackCounts"
          },
          "current_core_version": {
            "type": "string"
          },
          "defaults_will_adopt": {
            "type": "boolean"
          },
          "differences": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "installed_version": {
            "type": [
              "null",
              "string"
            ]
          },
          "integrity_digest": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "pack_id": {
            "type": "string"
          },
          "preview_id": {
            "format": "uuid",
            "type": "string"
          },
          "required_core_version": {
            "type": [
              "null",
              "string"
            ]
          },
          "resources": {
            "items": {
              "$ref": "#/components/schemas/AgentPackResourceView"
            },
            "type": "array"
          },
          "revision": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "action",
          "preview_id",
          "pack_id",
          "owner",
          "bundled_version",
          "installed_version",
          "integrity_digest",
          "revision",
          "expires_at",
          "counts",
          "resources",
          "defaults_will_adopt",
          "required_core_version",
          "current_core_version",
          "differences",
          "warnings"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TinadecCore.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/agent-packs/install-preview": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentPackEnvelopeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPackPreviewView"
                }
              }
            },
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Quoted numeric agent pack installation revision for conditional updates.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: invalid_agent_pack_manifest."
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_management_forbidden."
          },
          "409": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_owner_conflict, agent_pack_version_hash_conflict."
          },
          "422": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_incompatible, invalid_agent_pack_manifest."
          }
        },
        "summary": "Preview an agent pack installation or upgrade",
        "tags": [
          "AgentPacks"
        ]
      }
    },
    "/api/v1/agent-packs/{packId}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPackPurgeView"
                }
              }
            },
            "description": "OK"
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_management_forbidden."
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_not_found."
          },
          "412": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_revision_conflict."
          },
          "428": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: if_match_required."
          }
        },
        "summary": "Permanently delete an installed agent pack and the runs it produced",
        "tags": [
          "AgentPacks"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPackInstallationDetail"
                }
              }
            },
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Quoted numeric agent pack installation revision for conditional updates.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_not_found."
          }
        },
        "summary": "Get an installed agent pack",
        "tags": [
          "AgentPacks"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Unique key for this preview application. Replays with the same key and request converge to the stored receipt.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "Quoted numeric installation revision from the preview or pack ETag. Required for upgrades; omit for a first installation.",
            "in": "header",
            "name": "If-Match",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentPackApplyRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPackApplyView"
                }
              }
            },
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Quoted numeric agent pack installation revision for conditional updates.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPackApplyView"
                }
              }
            },
            "description": "Created",
            "headers": {
              "ETag": {
                "description": "Quoted numeric agent pack installation revision for conditional updates.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: invalid_agent_pack_manifest."
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_management_forbidden."
          },
          "409": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_owner_conflict, agent_pack_version_hash_conflict, agent_pack_resource_conflict."
          },
          "412": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_preview_stale, agent_pack_revision_conflict."
          },
          "422": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_incompatible, invalid_agent_pack_manifest."
          }
        },
        "summary": "Install or upgrade an agent pack",
        "tags": [
          "AgentPacks"
        ]
      }
    },
    "/api/v1/agent-packs/{packId}/adopt-defaults": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPackInstallationView"
                }
              }
            },
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Quoted numeric agent pack installation revision for conditional updates.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_management_forbidden."
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_not_found."
          },
          "409": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_not_active."
          }
        },
        "summary": "Make this pack\u0027s activation resources the workspace defaults",
        "tags": [
          "AgentPacks"
        ]
      }
    },
    "/api/v1/agent-packs/{packId}/disable": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPackInstallationView"
                }
              }
            },
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Quoted numeric agent pack installation revision for conditional updates.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_management_forbidden."
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_not_found."
          }
        },
        "summary": "Disable an installed agent pack without deleting it",
        "tags": [
          "AgentPacks"
        ]
      }
    },
    "/api/v1/agent-packs/{packId}/enable": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPackInstallationView"
                }
              }
            },
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Quoted numeric agent pack installation revision for conditional updates.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_management_forbidden."
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 problem details. Possible code values: agent_pack_not_found."
          }
        },
        "summary": "Enable an installed agent pack",
        "tags": [
          "AgentPacks"
        ]
      }
    },
...
    "/api/v1/agents/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "AgentConfigurationEndpoints"
        ]
      }
    },
...
    "/api/v1/tina-chat/participants/{id}/inbox/{messageId}/ack": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "messageId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
...
}